Split trait_transformer back into its own crate
