+++ trait-transformer/Cargo.toml
@@ -1,0 +1,30 @@
+# Copyright (c) 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 <LICENSE-APACHE or
+# https://www.apache.org/licenses/LICENSE-2.0> or the MIT license
+# <LICENSE-MIT or https://opensource.org/licenses/MIT>, at your
+# option. This file may not be copied, modified, or distributed
+# except according to those terms.
+
+[package]
+name = "trait-transformer"
+version = "0.0.0"
+description = "Utilities for working with impl traits in Rust"
+categories = ["asynchronous", "no-std", "rust-patterns"]
+keywords = ["async", "trait", "impl"]
+license.workspace = true
+repository.workspace = true
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+proc-macro2 = "1.0"
+quote = "1.0"
+syn = { version = "2.0", features = ["full"] }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["rt"] }
